+++ src/schema/inputs/NewMaxLiftInput.ts
@@ -1,10 +1,10 @@
-import { Field, InputType, Int } from 'type-graphql';
+import { Field, InputType, Float } from 'type-graphql';
 
 @InputType()
 export class NewMaxLiftInput {
   @Field()
   exercise: string;
 
-  @Field(() => Int)
+  @Field(() => Float)
   weight: number;
 }
